third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/1.1/form-action-blocked-when-target-cross-site-window-expected.txt

CONSOLE MESSAGE: line 26: submit event fired.
CONSOLE ERROR: line 30: Refused to send form data to 'http://127.0.0.1:8000/navigation/resources/form-target.pl' because it violates the following Content Security Policy directive: "form-action 'none'".

CONSOLE MESSAGE: line 35: securitypolicyviolation event fired.
CONSOLE MESSAGE: line 36: securitypolicyviolation_event.documentURI=http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/form-action-blocked-when-target-cross-site-window.html
CONSOLE MESSAGE: line 37: securitypolicyviolation_event.blockedURI=http://127.0.0.1:8000/navigation/resources/form-target.pl
CONSOLE MESSAGE: line 38: securitypolicyviolation_event.violatedDirective=form-action
 

Tests that blocking form actions works correctly. If this test passes, you will see a console error, and will not see a page indicating a form was POSTed.
